+++ app/src/main/java/com/alkisstam/taskbar/service/OverlayService.kt
@@ -244,10 +244,10 @@ class OverlayService : Service(), LifecycleOwner, ViewModelStoreOwner, SavedStat
         addPillView()
         addSearchView()
         addQuickStripView()
+        addMusicPanelView()
         addVolumeScrimView()
         addVolumePanelView()
         addBrightnessPanelView()
-        addMusicPanelView()
         if (!observersStarted) {
             observersStarted = true
             observePillPosition()